% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{anova_filter}
\alias{anova_filter}
\title{ANOVA filter}
\usage{
anova_filter(
  y,
  x,
  nfilter = NULL,
  p_cutoff = 0.05,
  type = c("index", "names", "full")
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{nfilter}{Number of predictors to return. If \code{NULL} all predictors with
p values < \code{p_cutoff} are returned.}

\item{p_cutoff}{p value cut-off}

\item{type}{Type of vector returned. Default "index" returns indices,
"names" returns predictor names, "full" returns a matrix of p values.}
}
\value{
Integer vector of indices of filtered parameters (type = "index") or
character vector of names (type = "names") of filtered parameters. If
\code{type} is \code{"full"} full output from \link[Rfast:ftests]{Rfast::ftests} is returned.
}
\description{
Simple univariate filter using anova (Welch's F-test) using the
Rfast package for speed.
}
