% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_filter.R
\name{bin_stat_filter}
\alias{bin_stat_filter}
\title{Univariate filter for binary classification}
\usage{
bin_stat_filter(
  y,
  x,
  force_vars = NULL,
  nfilter = NULL,
  p_cutoff = 0.05,
  rsq_cutoff = NULL,
  type = c("index", "names", "full"),
  ...
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix or dataframe of predictors}

\item{force_vars}{Vector of column names within \code{x} which are always retained
in the model (i.e. not filtered). Default \code{NULL} means all predictors will
be passed to \code{filterFUN}.}

\item{nfilter}{Number of predictors to return. If \code{NULL} all predictors with
p-values < \code{p_cutoff} are returned.}

\item{p_cutoff}{p value cut-off}

\item{rsq_cutoff}{r^2 cutoff for removing predictors due to collinearity.
Default \code{NULL} means no collinearity filtering. Predictors are ranked based
on t-test. If 2 or more predictors are collinear, the first ranked
predictor by t-test is retained, while the other collinear predictors are
removed. See \code{\link[=collinear]{collinear()}}.}

\item{type}{Type of vector returned. Default "index" returns indices, "names"
returns predictor names, "full" returns a list of p values.}

\item{...}{optional arguments, e.g. \code{rsq_method}: see \code{\link[=collinear]{collinear()}}.}
}
\value{
Integer vector of indices of filtered parameters (type = "index") or
character vector of names (type = "names") of filtered parameters in order
of t-test/chi-squared test p-value. If \code{type} is \code{"full"} full output is
returned containing a matrix of results from \link[Rfast:ttests]{Rfast::ttests} and a matrix
of \code{\link[=chisq.test]{chisq.test()}} results.
}
\description{
Univariate statistic filter for dataframes of predictors with mixed numeric
and categorical datatypes. Different statistical tests are used depending on
the data type of predictors: t-test for continuous data, chi-squared test for
categorical data.
}
