% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_filter.R
\name{boot_ttest}
\alias{boot_ttest}
\title{Bootstrap t-test filter}
\usage{
boot_ttest(y, x, B = 50, ...)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{B}{Number of times to bootstrap}

\item{...}{Optional arguments passed to \code{\link[=ttest_filter]{ttest_filter()}}}
}
\value{
Integer vector of indices of filtered parameters (\code{type = "index"})
or if \code{type = "full"} a matrix of rankings from each bootstrap is returned.
}
\description{
Randomly samples predictors and averages the ranking from \code{\link[=ttest_filter]{ttest_filter()}} to
give an ensemble measure of best predictors by bootstrapped t-test.
}
\seealso{
\code{\link[=boot_filter]{boot_filter()}}
}
