% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boxplot_model}
\alias{boxplot_model}
\title{Boxplot model predictors}
\usage{
boxplot_model(fit, x, cols = NULL, palette = "Dark 3", ...)
}
\arguments{
\item{fit}{"nestedcv" object}

\item{x}{matrix of predictors}

\item{cols}{colour scheme}

\item{palette}{palette name (one of \code{hcl.pals()}) which is passed to
\link{hcl.colors}}

\item{...}{other arguments passed to \link{boxplot}.}
}
\value{
No return value
}
\description{
Boxplots to show range of model predictors to identify exceptional predictors
with excessively low or high values.
}
\seealso{
\link{nestcv.glmnet}
}
