% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{calib}
\alias{calib}
\title{Calibration of binary \code{nestcv.glmnet} fitted object
Calibration of probabilities by Platt scaling}
\usage{
calib(fit)
}
\arguments{
\item{fit}{a \code{nestcv.glmnet} fitted object}
}
\value{
updated \code{nestcv.glmnet} fitted object with calibrated probabilities
in the output dataframe of\code{nestcv.glmnet} fitted object
under the column of "calib.p"
}
\description{
Calibration of binary \code{nestcv.glmnet} fitted object
Calibration of probabilities by Platt scaling
}
