% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib.R
\name{calib_plot}
\alias{calib_plot}
\title{Plotting calibrated p values from Platt scaling with original predicted p values}
\usage{
calib_plot(fit, nbin)
}
\arguments{
\item{fit}{a \code{nestcv.glmnet} fitted object}

\item{nbin}{the number of bins to create the plot}
}
\description{
A plot created using ggplot2.
Basis of code built from https://github.com/etlundquist/eRic/
}
