% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.glmnet.R
\name{coef.cva.glmnet}
\alias{coef.cva.glmnet}
\title{Extract coefficients from a cva.glmnet object}
\usage{
\method{coef}{cva.glmnet}(object, ...)
}
\arguments{
\item{object}{Fitted \code{cva.glmnet} object.}

\item{...}{Other arguments passed to \code{coef.glmnet()} e.g. \code{s} the value of
lambda at which coefficients are required.}
}
\description{
Extracts model coefficients from a fitted \code{\link[=cva.glmnet]{cva.glmnet()}} object.
}
