% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{coef.nestcv.glmnet}
\alias{coef.nestcv.glmnet}
\title{Extract coefficients from nestcv.glmnet object}
\usage{
\method{coef}{nestcv.glmnet}(object, s = object$mean_lambda, ...)
}
\arguments{
\item{object}{Object of class \code{"nestcv.glmnet"}}

\item{s}{Value of penalty parameter lambda. Default is the mean of lambdas
with lowest deviance across outer folds.}

\item{...}{Other arguments passed to \link{coef.glmnet}}
}
\description{
Extracts coefficients from the final fit of a \code{"nestcv.glmnet"} object.
}
