% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{combo_filter}
\alias{combo_filter}
\title{Combo filter}
\usage{
combo_filter(y, x, nfilter, return = "names", ...)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{nfilter}{Number of predictors to return, using 1/2 from \code{uni_filter}
and 1/2 from \code{relieff_filter}. Since \code{unique} is applied, the final number
returned may be less than \code{nfilter}.}

\item{return}{Type of output returned. Default "names" returns predictor
names, "full" returns full output.}

\item{...}{Optional arguments passed via \link{relieff_filter} to
\link[CORElearn:attrEval]{CORElearn::attrEval}}
}
\value{
Ordered vector of names of filtered parameters. If \code{return} is
\code{"full"} a list of 2 vectors of full output from both \link{uni_filter}
and \link{relieff_filter} is returned.
}
\description{
Filter combining univariate filtering and reliefF filtering in equal measure.
}
