% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{cv.rf}
\alias{cv.rf}
\title{Inner CV for random forest}
\usage{
cv.rf(
  y,
  x,
  nfolds = 10,
  repeats = 1,
  mtry = if (!is.null(y) && !is.factor(y)) max(floor(ncol(x)/3), 1) else
    floor(sqrt(ncol(x))),
  ...
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{nfolds}{Number of folds for cross-validation}

\item{repeats}{Number of repeats for repeated CV}

\item{mtry}{Vector of possible values of \code{mtry} to tune. \code{mtry} is the
number of predictors tried at each split, see \link{randomForest}}

\item{...}{other arguments passed to \link[caret:train]{caret::train} or \link{randomForest}}
}
\description{
Using caret.
}
