% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{cv_coefs}
\alias{cv_coefs}
\title{Extract coefficients from outer CV glmnet models}
\usage{
cv_coefs(x)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} fitted object}
}
\value{
matrix of coefficients from outer CV glmnet models. Coefficients for
variables which are not present in a particular outer CV fold model are set
to 0.
}
\description{
Extracts coefficients from outer CV glmnet models from a \code{nestcv.glmnet}
fitted object.
}
\seealso{
\code{\link[=cv_varImp]{cv_varImp()}}
}
