% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{cv_varImp}
\alias{cv_varImp}
\title{Extract coefficients from outer CV caret models}
\usage{
cv_varImp(x)
}
\arguments{
\item{x}{a \code{nestcv.train} fitted object}
}
\value{
matrix of variable importance from outer CV caret models. Variable
importance for variables which are not present in a particular outer CV
fold model is set to 0.
}
\description{
Extracts coefficients from outer CV glmnet models from a \code{nestcv.train}
fitted object.
}
\details{
Note that \code{\link[caret:varImp]{caret::varImp()}} may require the model package to be fully loaded
in order to function. During the fitting process \code{caret} often only loads the
package by namespace.
}
\seealso{
\link{cv_coefs}
}
