% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{double_cv}
\alias{double_cv}
\title{Performs double k-fold cross-validation, aggregating over many random splits of the data}
\usage{
double_cv(X, Y, funcs, reps = 50, n_folds = 10, alpha = 0.1, running = FALSE)
}
\arguments{
\item{X}{A nxp data matrix.}

\item{Y}{A n-vector of responses}

\item{funcs}{A list of containing three functions used as subroutines:
\describe{
  \item{\code{fitter}}{A function taking X and Y and returning a model fit object.}
  \item{\code{predictor}}{A function taking a model fit object from above
  and X' and returning a vector of predictions.}
  \item{\code{loss}}{A function taking a vector of predictions and true values returning
  a vector of losses.}
}}

\item{reps}{Number of repitions of double CV to combine. Many iterations are needed for stability.}

\item{n_folds}{Number of folds, an integer.}

\item{alpha}{Nominal type-I error level. Must be in (0,.5)}

\item{running}{Whether or not to compute estimate of inflation online, with automatic stopping.
Can be slow if automatic stopping is used.}
}
\value{
\describe{
  \item{\code{mean}}{Point estimate of prediction error: mean of CV prediction errors.}
  \item{\code{ci_lo}}{Lower endpoint of naive CV interval for prediction error.d}
  \item{\code{ci_hi}}{Lower endpoint of naive CV interval for prediction error.d}
  \item{\code{sd}}{Standard deviation of CV prediction errors.}
  \item{\code{raw_errors}}{Errors for each point, a vector of same length as input Y.}
  \item{\code{fold_id}}{The fold assignmnet for each point, a vector of same length as input Y.}
}
}
\description{
Performs double k-fold cross-validation, aggregating over many random splits of the data
}
