% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{double_cv_helper}
\alias{double_cv_helper}
\title{Internal helper for doubl_cv}
\usage{
double_cv_helper(X, Y, funcs, n_folds = 10)
}
\value{
\describe{
  \item{\code{pivots}}{A vector of same length as Y of difference statistics.}
  \item{\code{errors}}{A vector of all errors of observations not used in model training.}
}
}
\description{
arguments as in "double_cv" function
}
