% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{glmnet_coefs}
\alias{glmnet_coefs}
\title{glmnet coefficients}
\usage{
glmnet_coefs(fit, s)
}
\arguments{
\item{fit}{A \link{cv.glmnet} fitted model object.}

\item{s}{Value of lambda. See \link{coef.glmnet} and \link{predict.cv.glmnet}}
}
\value{
Vector of coefficients ordered with the intercept first, followed
by highest absolute value to lowest.
}
\description{
Convenience function for retrieving coefficients from a \link{cv.glmnet} model at
a specified lambda. Sparsity is removed and non-intercept coefficients are
ranked by absolute value.
}
