% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innercv_roc.R
\name{innercv_roc}
\alias{innercv_roc}
\alias{innercv_roc.nestcv.glmnet}
\alias{innercv_roc.nestcv.train}
\title{Build ROC curve from left-out folds from inner CV}
\usage{
innercv_roc(x, ...)

\method{innercv_roc}{nestcv.glmnet}(x, direction = "<", ...)

\method{innercv_roc}{nestcv.train}(x, direction = "<", ...)
}
\arguments{
\item{x}{Fitted \code{nestedcv} object}

\item{...}{Other arguments passed to \link[pROC:roc]{pROC::roc}}

\item{direction}{Set ROC directionality \link[pROC:roc]{pROC::roc}}
}
\value{
\code{"roc"} object, see \link[pROC:roc]{pROC::roc}
}
\description{
Build ROC (receiver operating characteristic) curve from left-out folds
from inner CV. Object can be plotted using \code{plot()} or passed to functions
\code{\link[=auc]{auc()}} etc.
}
