% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{innercv_roc2}
\alias{innercv_roc2}
\title{Build ROC curve from left-out folds from inner CV}
\usage{
innercv_roc2(x, direction = "<", ...)
}
\arguments{
\item{x}{Fitted \code{"nestcv.train"} object}

\item{direction}{Passed to \link[pROC:roc]{pROC::roc}}

\item{...}{Other arguments passed to \link[pROC:roc]{pROC::roc}}
}
\value{
\code{"roc"} object, see \link[pROC:roc]{pROC::roc}
}
\description{
Build ROC (receiver operating characteristic) curve from left-out folds
from inner CV. Object can be plotted or passed to functions \link{auc} etc.
}
