% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innercv_roc.R
\name{innercv_summary}
\alias{innercv_summary}
\title{Summarise performance on inner CV test folds}
\usage{
innercv_summary(x)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} object}
}
\value{
Returns performance metrics from outer training folds, see
\link{predSummary}.
}
\description{
Calculates performance metrics on inner CV held-out test folds: confusion
matrix, accuracy and balanced accuracy for classification; ROC AUC for binary
classification; RMSE for regression.
}
\details{
Note: currently only works for \code{nestcv.glmnet} objects.
}
\seealso{
\link{predSummary}
}
