% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_filters.R
\name{layer_filter}
\alias{layer_filter}
\title{Multilayer filter}
\usage{
layer_filter(
  y,
  x,
  nfilter = NULL,
  imbalance = TRUE,
  cull = 5,
  force_vars = NULL,
  verbose = FALSE,
  type = c("index", "names", "full")
)
}
\description{
Multilayer filter
}
