% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_recall.R
\name{lines.prc}
\alias{lines.prc}
\title{Add precision-recall curve to a plot}
\usage{
\method{lines}{prc}(x, ...)
}
\arguments{
\item{x}{An object of class 'prc'}

\item{...}{Optional graphical arguments passed to \code{\link[=lines]{lines()}}}
}
\value{
No return value
}
\description{
Adds a precision-recall curve to a base graphics plot. It accepts an S3
object of class 'prc', see \code{\link[=prc]{prc()}}.
}
\seealso{
\code{\link[=prc]{prc()}} \code{\link[=plot.prc]{plot.prc()}}
}
