% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_registerDoFuture}
\alias{local_registerDoFuture}
\title{Locally \code{registerDoFuture()} until end of the current function}
\usage{
local_registerDoFuture(.local_envir = parent.frame())
}
\arguments{
\item{.local_envir}{The environment to which exit handlers should be
attached. Users should generally not need to set this argument. See the
\code{envir} option of \code{\link[withr:defer]{withr::defer()}}.}
}
\description{
This function will call \code{\link[doFuture:registerDoFuture]{doFuture::registerDoFuture()}}, which will be in
effect until the end of the current function. After that, the originally
registered foreach backend will be restored.
}
\examples{
library(foreach)
library(future)

# This function will always use the registered future plan for
# parallelization, even if a different foreach backend is registered.
my_parallel_sqrt_function <- function(x) {
  local_registerDoFuture()
  foreach(x = x) \%dopar\% { sqrt(x) }
}
my_parallel_sqrt_function(1:10)
}
