% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.hsstan.R
\name{model.hsstan}
\alias{model.hsstan}
\title{hsstan model to be fitted with cross-validation}
\usage{
model.hsstan(y, x, unpenalized = NULL, ...)
}
\arguments{
\item{y}{Response vector. For classification this should be a factor.}

\item{x}{Matrix of predictors}

\item{unpenalized}{Vector of column names \code{x} which are always retained into
the model (i.e. not penalized). Default \code{NULL} means the parameters for all
predictors will be drawn from a hierarchical prior distribution, i.e. will
be penalized. Note: if filtering of predictors is specified, then the
vector of \code{unpenalized} predictors should also be passed to the filter
function using the \code{filter_options$force_vars} argument. Filters currently
implementing this option are the \code{partial_ttest_filter} for binary outcomes
and the \code{lm_filter} for continuous outcomes.}

\item{...}{Optional arguments passed to \code{hsstan}}
}
\value{
An object of class \code{hsstan}
}
\description{
This function applies a cross-validation (CV) procedure for training Bayesian
models with hierarchical shrinkage priors using the \code{hsstan} package.
The function allows the option of embedded filtering of predictors for
feature selection within the CV loop. Within each training fold, an optional
filtering of predictors is performed, followed by fitting of an \code{hsstsan}
model. Predictions on the testing folds are brought back together and error
estimation/ accuracy determined. The default is 10-fold CV.
The function is implemented within the \code{nestedcv} package. The \code{hsstan}
models do not require tuning of meta-parameters and therefore only a single
CV procedure is needed to evaluate performance. This is implemented using the
\code{outer} CV procedure in the \code{nestedcv} package.
}
\author{
Athina Spiliopoulou
}
