% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{naive_cv}
\alias{naive_cv}
\title{Cross-validation}
\usage{
naive_cv(X, Y, funcs, n_folds = 10, alpha = 0.1)
}
\arguments{
\item{X}{A nxp data matrix.}

\item{Y}{A n-vector of responses}

\item{funcs}{A list of containing three functions used as subroutines:
\describe{
  \item{\code{fitter}}{A function taking X and Y and returning a model fit object.}
  \item{\code{predictor}}{A function taking a model fit object from above
  and X' and returning a vector of predictions.}
  \item{\code{loss}}{A function taking a vector of predictions and true values returning
  a vector of losses.}
}}

\item{n_folds}{Number of folds, an integer.}

\item{alpha}{Nominal type-I error level. Must be in (0,.5)}
}
\value{
\describe{
  \item{\code{mean}}{Point estimate of prediction error: mean of CV prediction errors.}
  \item{\code{ci_lo}}{Lower endpoint of naive CV interval for prediction error.}
  \item{\code{ci_hi}}{Lower endpoint of naive CV interval for prediction error.}
  \item{\code{sd}}{Standard deviation of CV prediction errors.}
  \item{\code{raw_errors}}{Errors for each point, a vector of same length as input Y.}
  \item{\code{fold_id}}{The fold assignmnet for each point, a vector of same length as input Y.}
}
}
\description{
Performs standard k-fold cross-validation to estimate prediction
error and provide a confidence interval for it.
}
