% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nested_cv}
\alias{nested_cv}
\title{nested Cross-validation}
\usage{
nested_cv(
  X,
  Y,
  funcs,
  reps = 50,
  n_folds = 10,
  alpha = 0.1,
  bias_reps = NA,
  trans = list(identity),
  funcs_params = NULL,
  n_cores = 1
)
}
\arguments{
\item{X}{A nxp data matrix.}

\item{Y}{A n-vector of responses}

\item{funcs}{A list of containing three functions used as subroutines:
\describe{
  \item{\code{fitter}}{A function taking X and Y and returning a model fit object.}
  \item{\code{predictor}}{A function taking a model fit object from above
  and X' and returning a vector of predictions.}
  \item{\code{loss}}{A function taking a vector of predictions and true values returning
  a vector of losses.}
}}

\item{reps}{Number of repitions of nested CV to combine. Many iterations are needed for stability.}

\item{n_folds}{Number of folds, an integer.}

\item{alpha}{Nominal type-I error level. Must be in (0,.5)}
}
\value{
\describe{
  \item{\code{mean}}{Point estimate of prediction error: mean of CV prediction errors.}
  \item{\code{ci_lo}}{Lower endpoint of naive CV interval for prediction error.d}
  \item{\code{ci_hi}}{Lower endpoint of naive CV interval for prediction error.d}
  \item{\code{sd_infl}}{The multiplicative factor needed to correct the interval width.}
  \item{\code{sd}}{Standard deviation of CV prediction errors.}
  \item{\code{raw_errors}}{Errors for each point, a vector of same length as input Y.}
  \item{\code{fold_id}}{The fold assignmnet for each point, a vector of same length as input Y.}
  \item{\code{running_sd_infl}}{The value of "sd_infl" after each repitition. This is used
    to evaluate how stable the estimate is and if more replicates are needed.}
}
}
\description{
Performs nested k-fold cross-validation, aggregating over many random splits of the data.
Provides a confidence interval for prediction interval that is more accurate than that of
standard cross-validation.
}
