% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{nested_cv_helper}
\alias{nested_cv_helper}
\title{Internal helper for nested_cv}
\usage{
nested_cv_helper(
  X,
  Y,
  funcs,
  n_folds = 10,
  trans = list(identity),
  funcs_params = NULL
)
}
\value{
\describe{
  \item{\code{pivots}}{A vector of same length as Y of difference statistics.}
  \item{\code{errors}}{A vector of all errors of observations not used in model training.}
}
}
\description{
Randomly assigns folds and does one iteration of nested cross-validation.
}
\details{
arguments as in the "nested_cv" function
}
