% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedCV.R
\name{nestedCV}
\alias{nestedCV}
\title{nestedCV}
\usage{
nestedCV()
}
\arguments{
\item{data}{Dataframe containing all observations and features}
}
\value{
Nested cross validation accuracy over all outer folds
}
\description{
Description for function
}
