% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outercv.R
\name{outercv.rf}
\alias{outercv.rf}
\title{Outer cross-validation with random forest}
\usage{
outercv.rf(
  y,
  x,
  filterFUN = NULL,
  filter_options = NULL,
  n_outer_folds = 10,
  cores = 1,
  ...
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{filterFUN}{Filter function, e.g. \link{ttest_filter} or \link{relieff_filter}.
Any function can be provided and is passed \code{y} and \code{x}. Must return a
character vector with names of filtered predictors.}

\item{filter_options}{List of additional arguments passed to the filter
function specified by \code{filterFUN}.}

\item{n_outer_folds}{Number of outer CV folds}

\item{cores}{Number of cores for parallel processing. Note this currently
uses \link[parallel:mclapply]{parallel::mclapply}.}

\item{...}{Optional arguments passed to \link{randomForest}.}
}
\value{
An object with S3 class "outercv.rf"
}
\description{
Outer cross-validation (CV) with random forest. This is designed to evaluate
performance of the model with fixed hyperparameters and no tuning. If tuning
of parameters on data is required, this will need full nested CV with inner
CV to tune model hyperparameters (see \link{nestcv.train}).
}
