% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oversample.R
\name{oversample}
\alias{oversample}
\title{Oversampling and undersampling}
\usage{
oversample(y, over = NULL, under = 1, yminor = NULL)
}
\arguments{
\item{y}{Vector of response outcome as a factor}

\item{over}{Amount of oversampling of the minority class. If set to \code{NULL}
then all classes will be oversampled up to the number of samples in the
majority class.}

\item{under}{Amount of undersampling of the majority class}

\item{yminor}{Optional character value specifying the level in \code{y} which is
to be oversampled. If \code{NULL}, this is set automatically to the class with
the smallest sample size.}
}
\value{
A vector of sample indices
}
\description{
Random oversampling to compensate for data imbalance
}
