% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.cva.glmnet}
\alias{plot.cva.glmnet}
\title{Plot lambda across range of alphas}
\usage{
\method{plot}{cva.glmnet}(
  x,
  xaxis = c("lambda", "nvar"),
  errorBar = FALSE,
  errorWidth = 0.01,
  min.pch = NULL,
  cols = NULL,
  palette = "zissou",
  showLegend = "bottomright",
  ...
)
}
\arguments{
\item{x}{Object of class 'cva.glmnet'}

\item{xaxis}{String specifying what is plotted on the x axis, either log
lambda or the number of non-zero coefficients.}

\item{errorBar}{Logical whether to show error bars for the standard deviation.}

\item{errorWidth}{Width of error bars.}

\item{min.pch}{Plotting 'character' for the minimum point of each curve. Not
shown if set to \code{NULL}. See \link{points}}

\item{cols}{Colour scheme}

\item{palette}{Palette name (one of \code{hcl.pals()}) which is passed to
\link{hcl.colors}}

\item{showLegend}{Either a keyword to position the legend or \code{NULL} to hide
the legend.}

\item{...}{Other arguments passed to \link{plot}. Use \code{type = 'p'} to plot a
scatter plot instead of a line plot.}
}
\value{
No return value
}
\description{
Plot showing cross-validated tuning of alpha and lambda from elastic net
regression via \link{glmnet}. Log lambda is on the x axis while the tuning metric
(log loss, deviance, accuracy, AUC etc) is on the y axis. Multiple alpha
values are shown by different colours.
}
\seealso{
\link{nestcv.glmnet}
}
\author{
Myles Lewis
}
