% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_recall.R
\name{plot.prc}
\alias{plot.prc}
\title{Plot precision-recall curve}
\usage{
\method{plot}{prc}(x, ...)
}
\arguments{
\item{x}{An object of class 'prc'}

\item{...}{Optional graphical arguments passed to \code{\link[=plot]{plot()}}}
}
\value{
No return value
}
\description{
Plots a precision-recall curve using base graphics. It accepts an S3 object
of class 'prc', see \code{\link[=prc]{prc()}}.
}
\seealso{
\code{\link[=prc]{prc()}}
}
