% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{plot_alphas}
\alias{plot_alphas}
\title{Plot cross-validated glmnet alpha}
\usage{
plot_alphas(x, xlab = "Alpha", ylab = "Deviance", ...)
}
\arguments{
\item{x}{Fitted "nestcv.glmnet" object}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{...}{other arguments passed to plot}
}
\value{
No return value
}
\description{
Plot of cross-validated glmnet alpha parameter against deviance.
}
