% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_caret}
\alias{plot_caret}
\title{Plot caret tuning}
\usage{
plot_caret(x, ...)
}
\arguments{
\item{x}{Object of class 'train' generated by \code{caret} function \link{train}}

\item{...}{Other arguments passed to \code{\link[=plot]{plot()}}}
}
\description{
Plots the main tuning parameter in models built using \link[caret:train]{caret::train}
}
