% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lambdas}
\alias{plot_lambdas}
\title{Plot cross-validated glmnet lambdas across outer folds}
\usage{
plot_lambdas(x, col = NULL, ...)
}
\arguments{
\item{x}{Fitted "nestcv.glmnet" object}

\item{col}{Optional vector of line colours for each fold}

\item{...}{other arguments passed to plot}
}
\value{
No return value
}
\description{
Plot of cross-validated glmnet lambda parameter against deviance for each
outer CV fold.
}
