% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap.R
\name{plot_shap_importance}
\alias{plot_shap_importance}
\title{SHAP importance beeswarm plot}
\usage{
plot_shap_importance(shap, x, bee.cex = 0.5)
}
\arguments{
\item{shap}{a matrix of SHAP values}

\item{x}{a matrix of feature values containing only features values from the
training data. The rows must match rows in \code{shap}.}

\item{bee.cex}{Scaling for adjusting point spacing. See \code{cex} in
\code{ggbeeswarm::geom_beeswarm()}.}
}
\description{
SHAP importance beeswarm plot
}
