% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{plot_var_stability}
\alias{plot_var_stability}
\title{Plot variable stability}
\usage{
plot_var_stability(
  x,
  final = TRUE,
  top = 25,
  direction = 0,
  dir_labels = NULL,
  percent = TRUE,
  breaks = c(2, 4, 6, 8, 10)
)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} or \code{nestcv.train} fitted object}

\item{final}{Logical whether to restrict variables to only those which ended
up in the final fitted model or to include all variables selected across
all outer folds.}

\item{top}{Limits number of variables plotted. Ignored if \code{final = TRUE}.}

\item{direction}{Integer controlling plotting of directionality for binary or
regression models. 0 means no directionality is shown, 1 means
directionality is overlaid as a colour, 2 means directionality is reflected
in the sign of variable importance.}

\item{dir_labels}{Character vector for controlling the legend when
\code{direction = 1}}

\item{percent}{Logical for \code{nestcv.glmnet} objects only, whether to scale
coefficients to percentage of the largest coefficient in each model. If set
to \code{FALSE}, model coefficients are shown and \code{direction} is ignored.}

\item{breaks}{Vector of continuous breaks for legend colour/size}
}
\value{
A ggplot2 plot
}
\description{
Produces a ggplot2 plot of stability (as SEM) of variable importance across
models trained and tested across outer CV folds. Overlays frequency with
which variables are selected across the outer folds and optionally overlays
directionality for binary response outcome.
}
\seealso{
\code{\link[=var_stability]{var_stability()}}
}
