% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_varImp}
\alias{plot_varImp}
\title{Variable importance plot}
\usage{
plot_varImp(x, abs = TRUE, size = TRUE)
}
\arguments{
\item{x}{a 'nestcv.glmnet' class object}

\item{abs}{Logical whether to show absolute value of glmnet coefficients}

\item{size}{Logical whether to show mean expression by size of points}
}
\value{
Returns a ggplot2 plot
}
\description{
Plot of variable importance of coefficients of a final fitted
'nestedcv.glmnet' model using ggplot2. Mean expression can be overlaid as the
size of points as this can be informative in models of biological attributes.
}
