% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.glmnet.R
\name{predict.cva.glmnet}
\alias{predict.cva.glmnet}
\title{Predict method for cva.glmnet models}
\usage{
\method{predict}{cva.glmnet}(object, newx, s = "lambda.1se", ...)
}
\arguments{
\item{object}{Fitted \code{cva.glmnet} object.}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be made.}

\item{s}{Value of penalty parameter lambda. Default value is \code{s="lambda.1se"}
for consistency with glmnet. Alternatively \code{s="lambda.min"} can be used.}

\item{...}{Other arguments passed to \code{predict.cv.glmnet()}.}
}
\value{
Object returned depends on arguments in \code{...} such as \code{type}.
}
\description{
Makes predictions from a cross-validated glmnet model with optimal value of
lambda and alpha.
}
