% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSummary.R
\name{predSummary}
\alias{predSummary}
\title{Summarise prediction performance metrics}
\usage{
predSummary(output)
}
\arguments{
\item{output}{data.frame with columns \code{testy} containing observed response
from test folds; \code{predy} predicted response; \code{predyp} (optional) predicted
probabilities for classification to calculate ROC AUC}
}
\value{
Vector containing accuracy and balanced accuracy for classification,
ROC AUC for binary classification, RMSE for regression.
}
\description{
Quick function to calculate performance metrics: accuracy and balanced
accuracy for classification; ROC AUC for binary classification; RMSE for
regression.
}
