% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess data}
\usage{
preprocess(
  y,
  x,
  balance = c("none", "randomsample", "smote"),
  balance_options = list(minor = NULL, major = 1),
  na.option = "pass"
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{balance}{Specifies method for dealing with imbalanced data. Current
options are \code{"none"} or \code{"randomsample"}. See \code{\link[=randomsample]{randomsample()}}}

\item{balance_options}{List of additional arguments passed to the balancing
function}

\item{na.option}{Character value specifying how \code{NA}s are dealt with.
\code{"omit"} (the default) is equivalent to \code{na.action = na.omit}. \code{"omitcol"}
removes cases if there are \code{NA} in 'y', but columns (predictors) containing
\code{NA} are removed from 'x' to preserve cases. Any other value means that
\code{NA} are ignored (a message is given).}
}
\value{
A list containing \code{y}, \code{x} and \code{pre} which is a list containing
balancing information and original values of \code{y}
}
\description{
Preprocessing of datasets including checking and optionally omitting \code{NA} and
handling of class imbalance.
}
