% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{relieff_filter}
\alias{relieff_filter}
\title{ReliefF filter}
\usage{
relieff_filter(
  y,
  x,
  nfilter = NULL,
  estimator = "ReliefFequalK",
  return = "names",
  ...
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{nfilter}{Number of predictors to return. If \code{NULL} all predictors are
returned.}

\item{estimator}{Type of algorithm used, see \link[CORElearn:attrEval]{CORElearn::attrEval}}

\item{return}{Type of vector returned. Default "names" returns predictor
names, "full" returns a named vector of variable importance.}

\item{...}{Other arguments passed to \link[CORElearn:attrEval]{CORElearn::attrEval}}
}
\value{
Ordered vector of names of filtered parameters. If \code{return} is
\code{"full"} a named vector of variable importance is returned.
}
\description{
Uses ReliefF algorithm from the CORElearn package to rank predictors in order
of importance.
}
