% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestcv.train.R
\name{summary_vars}
\alias{summary_vars}
\title{Summarise variables}
\usage{
summary_vars(x)
}
\arguments{
\item{x}{Matrix or dataframe with variables in columns}
}
\value{
A matrix with variables in rows and mean, median and SD for each
variable or number of levels if the variable is a factor. If \code{NA} are
detected, an extra column \code{n.NA} is added with the numbers of \code{NA} for each
variable.
}
\description{
Summarise variables
}
