% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestcv.train.R
\name{swapFoldIndex}
\alias{swapFoldIndex}
\title{Convert testing folds to training folds and \emph{vice versa}}
\usage{
swapFoldIndex(folds, len = max(unlist(folds)))
}
\arguments{
\item{folds}{List of folds such as that created by \code{\link[caret:createDataPartition]{caret::createFolds()}}}

\item{len}{Number of observations in the data. This is only required if not
all indices are present in \code{folds}.}
}
\value{
A list the same length as \code{folds}, with the indices inverted: if the
input was test folds, then the output is training folds, and \emph{vice versa}.
}
\description{
Convert testing folds to training folds and \emph{vice versa}
}
\examples{
set.seed(1986)
test_folds <- caret::createFolds(1:10, k = 3)
train_folds <- swapFoldIndex(test_folds)
lengths(test_folds)
lengths(train_folds)
# Together, they add up to all 10 samples
lengths(test_folds) + lengths(train_folds)
}
