% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innercv_roc.R
\name{train_preds}
\alias{train_preds}
\title{Outer training fold predictions}
\usage{
train_preds(x)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} object}
}
\value{
Dataframe with columns \code{ytrain} and \code{predy} containing observed and
predicted values from training folds. For binomial and multinomial models
additional columns are added with class probabilities or log likelihood
values.
}
\description{
Obtain predictions on outer training folds which can be used for performance
metrics and ROC curves.
}
