% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innercv_roc.R
\name{train_roc}
\alias{train_roc}
\title{Build ROC curve from outer CV training folds}
\usage{
train_roc(x, direction = "<", ...)
}
\arguments{
\item{x}{Fitted \code{nestcv.glmnet} object}

\item{direction}{Set ROC directionality \link[pROC:roc]{pROC::roc}}

\item{...}{Other arguments passed to \link[pROC:roc]{pROC::roc}}
}
\value{
\code{"roc"} object, see \link[pROC:roc]{pROC::roc}
}
\description{
Build ROC (receiver operating characteristic) curve from outer training
folds. Object can be plotted using \code{plot()} or passed to functions \code{\link[=auc]{auc()}}
etc.
}
\details{
Note: currently only works for \code{nestcv.glmnet} objects. The argument
\code{outer_train_predict} must be set to \code{TRUE} in the original nested CV call.
}
