% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtProgressBar2.R
\name{txtProgressBar2}
\alias{txtProgressBar2}
\title{Text Progress Bar 2}
\usage{
txtProgressBar2(
  min = 0,
  max = 1,
  initial = 0,
  char = "=",
  width = NA,
  title = ""
)
}
\arguments{
\item{min}{Numeric value for minimum of the progress bar.}

\item{max}{Numeric value for maximum of the progress bar.}

\item{initial}{Initial value for the progress bar.}

\item{char}{The character (or character string) to form the progress bar.}

\item{width}{The width of the progress bar, as a multiple of the width of
\code{char}. If \code{NA}, the default, the number of characters is that which fits
into \code{getOption("width")}.}

\item{title}{Title for the progress bar.}
}
\value{
An object of class "\code{txtProgressBar}".
}
\description{
Text progress bar in the R console. Modified from \code{utils::txtProgressBar()}
to include title and timing.
}
\details{
Use \code{utils::setTxtProgressBar()} to set the progress bar and \code{close()} to
close it.
}
