% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{uni_filter}
\alias{uni_filter}
\title{Univariate filter}
\usage{
uni_filter(y, x, nfilter = NULL, p_cutoff = 0.05, return = "names")
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{nfilter}{Number of predictors to return. If \code{NULL} all predictors with
p values < \code{p_cutoff} are returned.}

\item{p_cutoff}{p value cut-off}

\item{return}{Type of vector returned. Default "names" returns predictor
names, "full" returns a named vector of p values.}
}
\value{
Ordered vector of names of filtered parameters. If \code{return} is
\code{"full"} a named vector of p values is returned.
}
\description{
Simple univariate filter using t-test or anova.
}
