% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionality.R
\name{var_direction}
\alias{var_direction}
\alias{var_direction.default}
\alias{var_direction.nestcv.train}
\title{Directionality of final predictors}
\usage{
var_direction(object)

\method{var_direction}{default}(object)

\method{var_direction}{nestcv.train}(object)
}
\description{
Directionality of final predictors
}
