% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{var_stability}
\alias{var_stability}
\alias{var_stability.nestcv.glmnet}
\alias{var_stability.nestcv.train}
\title{Variable stability}
\usage{
var_stability(x, ...)

\method{var_stability}{nestcv.glmnet}(x, percent = FALSE, level = 1, sort = TRUE, ...)

\method{var_stability}{nestcv.train}(x, ...)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} or \code{nestcv.train} fitted object}

\item{...}{Optional arguments for compatibility}

\item{percent}{Logical for \code{nestcv.glmnet} objects only, whether to scale
coefficients to percentage of the largest coefficient in each model}

\item{level}{For multinomial \code{nestcv.glmnet} models only, either an integer
specifying which level of outcome is being examined, or the level can be
specified as a character value}

\item{sort}{Logical whether to sort variables by mean importance}
}
\value{
Dataframe containing mean, sd, sem of variable importance and
frequency by which each variable is selected in outer folds.
}
\description{
Uses variable importance across models trained and tested across outer CV
folds to assess stability of variable importance.
}
\details{
Note that for caret models \code{\link[caret:varImp]{caret::varImp()}} may require the model package to
be fully loaded in order to function. During the fitting process \code{caret}
often only loads the package by namespace.
}
\seealso{
\code{\link[=cv_coef]{cv_coef()}} \code{\link[=cv_varImp]{cv_varImp()}}
}
