% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_filter_balance.R
\name{weight}
\alias{weight}
\title{Calculate weights for class imbalance}
\usage{
weight(y)
}
\arguments{
\item{y}{Factor or character response vector. If a character vector is
supplied it is coerced into a factor. Unused levels are dropped.}
}
\value{
Vector of weights
}
\description{
Calculate weights for class imbalance
}
