% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_balance.R
\name{class_balance}
\alias{class_balance}
\alias{class_balance.default}
\alias{class_balance.nestcv.train}
\title{Check class balance in training folds}
\usage{
class_balance(object)

\method{class_balance}{default}(object)

\method{class_balance}{nestcv.train}(object)
}
\arguments{
\item{object}{Object of class \code{nestedcv.glmnet}, \code{nestcv.train} or \code{outercv}}
}
\value{
Invisibly a table of the response classes in the training folds
}
\description{
Check class balance in training folds
}
