% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.glmnet.R
\name{glmnet_coefs}
\alias{glmnet_coefs}
\title{glmnet coefficients}
\usage{
glmnet_coefs(fit, s, ...)
}
\arguments{
\item{fit}{A \link[glmnet:cv.glmnet]{glmnet::cv.glmnet} fitted model object.}

\item{s}{Value of lambda. See \link[glmnet:predict.glmnet]{glmnet::coef.glmnet} and
\link[glmnet:predict.cv.glmnet]{glmnet::predict.cv.glmnet}}

\item{...}{Other arguments passed to \link[glmnet:predict.glmnet]{glmnet::coef.glmnet}}
}
\value{
Vector or list of coefficients ordered with the intercept first,
followed by highest absolute value to lowest.
}
\description{
Convenience function for retrieving coefficients from a \link[glmnet:cv.glmnet]{glmnet::cv.glmnet}
model at a specified lambda. Sparsity is removed and non-intercept
coefficients are ranked by absolute value.
}
