% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innercv_roc.R
\name{innercv_preds}
\alias{innercv_preds}
\alias{innercv_preds.nestcv.glmnet}
\alias{innercv_preds.nestcv.train}
\title{Inner CV predictions}
\usage{
innercv_preds(x)

\method{innercv_preds}{nestcv.glmnet}(x)

\method{innercv_preds}{nestcv.train}(x)
}
\arguments{
\item{x}{Fitted \code{nestedcv} object}
}
\value{
Dataframe with columns \code{testy} and \code{predy}, and for binomial and
multinomial models additional columns containing probabilities or log
likelihood values.
}
\description{
Obtain predictions on held-out test inner CV folds
}
