% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot}
\alias{one_hot}
\title{One-hot encode}
\usage{
one_hot(x, all_levels = TRUE, rename_binary = TRUE, sep = ".")
}
\arguments{
\item{x}{A dataframe or matrix. Matrices are returned untouched.}

\item{all_levels}{Logical, whether to create dummy variables for all levels
of each factor.}

\item{rename_binary}{Logical, whether to rename binary factors by appending
the 2nd level of the factor to aid interpretation of encoded factor levels
and to allow consistency with naming.}

\item{sep}{Character for separating factor variable names and levels for
encoded columns.}
}
\value{
A numeric matrix with multi-level factors converted to one-hot
encoded extra columns with 0 and 1. Binary factors are each converted to
single columns of 0 and 1. Ordered factors are converted to integer levels.
}
\description{
One-hot encoding of factors in dataframes. Binary factors are each converted
to a single column of 0 and 1. Multi-level unordered factors are converted to
multiple columns, one for each level (dummy variables). Unused levels are
dropped. Numeric or integer columns are left untouched.
}
\details{
\code{all_levels} is set to \code{TRUE} by default to aid with interpretability e.g.
with SHAP values, and since filtering might result in some dummy variables
being excluded. However, having dummy variables for all levels of a factor
can cause problems with multicollinearity in regression (the dummy variable
trap), so for regression models such as \code{glmnet}, \code{all_levels} should be set
to \code{FALSE} (equivalent to full rank parameterisation). Note this function is
designed to quickly generate dummy variables for more general machine
learning purposes; it is not designed for use with a standard linear model
or GLM in R. To create a proper design matrix object for regression models,
\code{\link[=model.matrix]{model.matrix()}} should be used instead.
}
\examples{
data(iris)
x <- iris
x2 <- one_hot(x)
head(x2)  # 3 columns for Species

x2 <- one_hot(x, all_levels = FALSE)
head(x2)  # 2 columns for Species

}
\seealso{
\code{\link[caret:dummyVars]{caret::dummyVars()}}, \code{\link[=model.matrix]{model.matrix()}}
}
