% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outercv.R
\name{outercv}
\alias{outercv}
\title{Outer cross-validation of models}
\usage{
outercv(
  model,
  y,
  x,
  filterFUN = NULL,
  filter_options = NULL,
  n_outer_folds = 10,
  cores = 1,
  ...
)
}
\arguments{
\item{model}{String specifying function to fit. Currently supports
randomForest and naive_bayes.}

\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{filterFUN}{Filter function, e.g. \link{ttest_filter} or \link{relieff_filter}.
Any function can be provided and is passed \code{y} and \code{x}. Must return a
character vector with names of filtered predictors.}

\item{filter_options}{List of additional arguments passed to the filter
function specified by \code{filterFUN}.}

\item{n_outer_folds}{Number of outer CV folds}

\item{cores}{Number of cores for parallel processing. Note this currently
uses \link[parallel:mclapply]{parallel::mclapply}.}

\item{...}{Optional arguments passed to the function specified by \code{model}.}
}
\value{
An object with S3 class "outercv.rf"
}
\description{
Outer cross-validation (CV) of selected models. This is designed to quickly
evaluate performance of specific models with fixed hyperparameters and no
tuning. If tuning of parameters on data is required, full nested CV with
inner CV is recommended to tune model hyperparameters (see \link{nestcv.train}).
}
