% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outercv.R
\name{outercv}
\alias{outercv}
\title{Outer cross-validation of selected models}
\usage{
outercv(
  model,
  y,
  x,
  filterFUN = NULL,
  filter_options = NULL,
  n_outer_folds = 10,
  cores = 1,
  ...
)
}
\arguments{
\item{model}{String specifying function to fit. Currently supports
randomForest and naive_bayes.}

\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{filterFUN}{Filter function, e.g. \link{ttest_filter} or \link{relieff_filter}.
Any function can be provided and is passed \code{y} and \code{x}. Must return a
character vector with names of filtered predictors.}

\item{filter_options}{List of additional arguments passed to the filter
function specified by \code{filterFUN}.}

\item{n_outer_folds}{Number of outer CV folds}

\item{cores}{Number of cores for parallel processing. Note this currently
uses \link[parallel:mclapply]{parallel::mclapply}.}

\item{...}{Optional arguments passed to the function specified by \code{model}.}
}
\value{
An object with S3 class "outercv"
}
\description{
This is a convenience function designed to quickly evaluate performance of
specific models (random forest, naive Bayes) with fixed hyperparameters and
no tuning. If tuning of parameters on data is required, full nested CV with
inner CV is recommended to tune model hyperparameters (see \link{nestcv.train}).
}
\details{
An alternative method of tuning a single model with fixed parameters
is to use \link{nestcv.train} with \code{tuneGrid} set as a single row of a
data.frame. The parameters which are needed for a specific model can be
identified using \code{\link[caret:modelLookup]{caret::modelLookup()}}.
}
