% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{outercv.rf}
\alias{outercv.rf}
\title{Outer cross-validation with randomForest}
\usage{
outercv.rf(y, x, filterFUN = NULL, n_outer_folds = 10, cores = 1, ...)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{filterFUN}{Filter function, e.g. \link{uni_filter} or \link{relieff_filter}.
Any function can be provided and is passed \code{y} and \code{x}. Must return a
character vector with names of filtered predictors.}

\item{n_outer_folds}{Number of outer CV folds}

\item{cores}{Number of cores for parallel processing. Note this currently
uses \link[parallel:mclapply]{parallel::mclapply}.}

\item{...}{Optional arguments passed both to \link{randomForest} as well as the
filter function specified by \code{filterFUN}}
}
\value{
An object with S3 class "outercv.rf"
}
\description{
Outer cross-validation (CV) with randomForest. Note, no tuning of
parameters is performed. If tuning of parameters is required, use nestedcv.
}
