% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.cva.glmnet}
\alias{plot.cva.glmnet}
\title{Plot lambda across range of alphas}
\usage{
\method{plot}{cva.glmnet}(x, cols = NULL, palette = "Dark 3", showLegend = "topright", ...)
}
\arguments{
\item{x}{Object of class 'cva.glmnet'}

\item{cols}{colour scheme}

\item{palette}{palette name (one of \code{hcl.pals()}) which is passed to
\link{hcl.colors}}

\item{showLegend}{Either a keyword to position the legend or \code{NULL} to hide
the legend.}

\item{...}{Other arguments passed to \link{plot}. Use \code{type = 'p'} to plot a
scatter plot instead of a line plot.}
}
\value{
No return value
}
\description{
Plot showing cross-validated tuning of alpha and lambda from elastic net
regression via \link{glmnet}. Log lambda is on the x axis while the tuning metric
(log loss, deviance, accuracy, AUC etc) is on the y axis. Multiple alpha
values are shown by different colours.
}
\seealso{
\link{nestcv.glmnet}
}
