% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{plot_alphas}
\alias{plot_alphas}
\title{Plot cross-validated glmnet alpha}
\usage{
plot_alphas(x, col = NULL, ...)
}
\arguments{
\item{x}{Fitted "nestcv.glmnet" object}

\item{col}{Optional vector of line colours for each fold}

\item{...}{other arguments passed to plot}
}
\value{
No return value
}
\description{
Plot of cross-validated glmnet alpha parameter against deviance.
}
