% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lambdas}
\alias{plot_lambdas}
\title{Plot cross-validated glmnet lambdas across outer folds}
\usage{
plot_lambdas(
  x,
  cols = NULL,
  palette = "Dark 3",
  showLegend = if (x$outer_method == "cv") "topright" else NULL,
  ...
)
}
\arguments{
\item{x}{Fitted "nestcv.glmnet" object}

\item{cols}{colour scheme}

\item{palette}{palette name (one of \code{hcl.pals()}) which is passed to
\link{hcl.colors}}

\item{showLegend}{Either a keyword to position the legend or \code{NULL} to hide
the legend.}

\item{...}{other arguments passed to plot. Use \code{type = 'p'} to plot a
scatter plot instead of a line plot.}
}
\value{
No return value
}
\description{
Plot of cross-validated glmnet lambda parameter against deviance for each
outer CV fold.
}
\seealso{
\link{nestcv.glmnet}
}
