% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{plot_var_stability}
\alias{plot_var_stability}
\title{Plot variable stability}
\usage{
plot_var_stability(
  x,
  abs = TRUE,
  final = TRUE,
  top = 25,
  breaks = c(1, 2, 5, 10),
  percent = TRUE
)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} or \code{nestcv.train} fitted object}

\item{abs}{Logical for \code{nestcv.glmnet} objects only, whether to show absolute
coefficients}

\item{final}{Logical whether to restrict variables to only those which ended
up in the final fitted model or to include all variables selected across
all outer folds.}

\item{top}{Limits number of variables plotted. Ignored if \code{final = TRUE}.}

\item{breaks}{Vector of continuous breaks for legend colour/size}

\item{percent}{Logical for \code{nestcv.glmnet} objects only, whether to scale
coefficients to percentage of the largest coefficient in each model}
}
\value{
A ggplot2 plot
}
\description{
Produces a ggplot2 plot of stability of variable importance across models trained and
tested across outer CV folds.
}
\seealso{
\link{var_stability}
}
