% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImp.R
\name{plot_var_stability}
\alias{plot_var_stability}
\title{Plot variable stability}
\usage{
plot_var_stability(
  x,
  final = TRUE,
  top = 25,
  direction = 0,
  percent = TRUE,
  breaks = c(2, 4, 6, 8, 10)
)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} or \code{nestcv.train} fitted object}

\item{final}{Logical whether to restrict variables to only those which ended
up in the final fitted model or to include all variables selected across
all outer folds.}

\item{top}{Limits number of variables plotted. Ignored if \code{final = TRUE}.}

\item{direction}{Integer controlling plotting of directionality for binary
models. 0 means no directionality is shown, 1 means directionality is
overlaid as a colour, 2 means directionality is reflected in the sign of
variable importance.}

\item{percent}{Logical for \code{nestcv.glmnet} objects only, whether to scale
coefficients to percentage of the largest coefficient in each model. If set
to \code{FALSE}, model coefficients are shown and \code{direction} is ignored.}

\item{breaks}{Vector of continuous breaks for legend colour/size}
}
\value{
A ggplot2 plot
}
\description{
Produces a ggplot2 plot of stability (as SEM) of variable importance across
models trained and tested across outer CV folds. Overlays frequency with
which variables are selected across the outer folds and optionally overlays
directionality for binary response outcome.
}
\details{
When plotting variable stability for models trained using \code{caret} via
\code{\link[=nestcv.train]{nestcv.train()}}, additional packages may need to be loaded. Some
\code{\link[=nestcv.train]{nestcv.train()}} fitted objects give an error message due to the package
(e.g. \code{gbm}) for the model not being loaded when \code{\link[caret:varImp]{caret::varImp()}} is called
to obtain the variable importance from the fitted models.
}
\seealso{
\code{\link[=var_stability]{var_stability()}}
}
