% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_recall.R
\name{prc}
\alias{prc}
\alias{prc.default}
\alias{prc.data.frame}
\alias{prc.nestcv.glmnet}
\alias{prc.nestcv.train}
\alias{prc.nestcv.SuperLearner}
\alias{prc.outercv}
\alias{prc.repeatcv}
\title{Build precision-recall curve}
\usage{
prc(...)

\method{prc}{default}(response, predictor, ...)

\method{prc}{data.frame}(output, ...)

\method{prc}{nestcv.glmnet}(object, ...)

\method{prc}{nestcv.train}(object, ...)

\method{prc}{nestcv.SuperLearner}(object, ...)

\method{prc}{outercv}(object, ...)

\method{prc}{repeatcv}(object, ...)
}
\arguments{
\item{...}{other arguments}

\item{response}{binary factor vector of response of default order controls,
cases.}

\item{predictor}{numeric vector of probabilities}

\item{output}{data.frame with columns \code{testy} containing observed response
from test folds, and \code{predyp} predicted probabilities for classification}

\item{object}{a 'nestcv.glmnet', 'nestcv.train', 'nestcv.SuperLearn',
'outercv' or 'repeatcv' S3 class results object.}
}
\value{
An object of S3 class 'prc' containing the following fields:
\item{recall}{vector of recall values}
\item{precision}{vector of precision values}
\item{auc}{area under precision-recall curve value using trapezoid method}
}
\description{
Builds a precision-recall curve for a nestedcv model using \code{prediction()} and
\code{performance()} functions from the ROCR package and returns an object of
class 'prc' for plotting.
}
