% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap.R
\name{pred_nestcv_glmnet}
\alias{pred_nestcv_glmnet}
\alias{pred_train}
\title{Prediction wrappers to use fastshap with nestedcv}
\usage{
pred_nestcv_glmnet(x, newdata)

pred_train(x, newdata)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} or \code{nestcv.train} object}

\item{newdata}{a matrix of new data}
}
\value{
prediction wrapper function designed for use with
\code{\link[fastshap:explain]{fastshap::explain()}}
}
\description{
Prediction wrapper functions to enable the use of the \code{fastshap} package for
generating SHAP values from \code{nestedcv} trained models.
}
