% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.glmnet.R
\name{predict.nestcv.glmnet}
\alias{predict.nestcv.glmnet}
\title{Predict method for nestcv.glmnet fits}
\usage{
\method{predict}{nestcv.glmnet}(object, newdata, s = object$final_param["lambda"], ...)
}
\arguments{
\item{object}{Fitted \code{nestcv.glmnet} object}

\item{newdata}{New data to predict outcome on}

\item{s}{Value of lambda for glmnet prediction}

\item{...}{Other arguments passed to \code{predict.glmnet}.}
}
\value{
Object returned depends on the \code{...} argument passed to predict
method for \code{glmnet} objects.
}
\description{
Obtains predictions from the final fitted model from a \link{nestcv.glmnet}
object.
}
\details{
Checks for missing predictors and if these are sparse (i.e. have
zero coefficients) columns of 0 are automatically added to enable
prediction to proceed.
}
\seealso{
\link[glmnet:glmnet]{glmnet::glmnet}
}
