% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatcv.R
\name{repeatcv}
\alias{repeatcv}
\title{Repeated nested CV}
\usage{
repeatcv(expr, n = 5, repeat_folds = NULL, progress = TRUE)
}
\arguments{
\item{expr}{An expression containing a call to \code{\link[=nestcv.glmnet]{nestcv.glmnet()}},
\code{\link[=nestcv.train]{nestcv.train()}}, \code{\link[=nestcv.SuperLearner]{nestcv.SuperLearner()}} or \code{\link[=outercv]{outercv()}}.}

\item{n}{Number of repeats}

\item{repeat_folds}{Optional list containing fold indices to be applied to
the outer CV folds.}

\item{progress}{Logical whether to show progress.}
}
\value{
Matrix of performance metrics
}
\description{
Performs repeated calls to a \code{nestedcv} model to determine performance across
repeated runs of nested CV.
}
\details{
When comparing models, it is recommended to fix the sets of outer CV folds
used across each repeat for comparing performance between models. The
function \code{\link[=repeatfolds]{repeatfolds()}} can be used to create a fixed set of outer CV folds
for each repeat.
}
\examples{
\donttest{
library(mlbench)
data(BostonHousing2)
dat <- BostonHousing2
y <- dat$cmedv
x <- subset(dat, select = -c(cmedv, medv, town, chas))

res <- repeatcv(n = 3, nestcv.glmnet(y, x, family = "gaussian", alphaSet = 1,
                                     n_outer_folds = 4, cv.cores = 2))
res

## using magrittr nested pipe
`\%|>\%` <- magrittr::pipe_nested
res <- nestcv.glmnet(y, x, family = "gaussian", alphaSet = 1,
                     n_outer_folds = 4, cv.cores = 2) \%|>\%
       repeatcv(3)
res

## set up fixed fold indices
folds <- repeatfolds(y, repeats = 3, n_outer_folds = 4)
res <- nestcv.glmnet(y, x, family = "gaussian", alphaSet = 1,
                     n_outer_folds = 4, cv.cores = 2) \%|>\%
       repeatcv(3, repeat_folds = folds)
res
}
}
