% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedcv.R
\name{rf_filter}
\alias{rf_filter}
\title{Random forest filter}
\usage{
rf_filter(
  y,
  x,
  nfilter = NULL,
  return = "names",
  ntree = 1000,
  mtry = ncol(x) * 0.2,
  ...
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{nfilter}{Number of predictors to return. If \code{NULL} all predictors are
returned.}

\item{return}{Type of vector returned. Default "names" returns predictor
names, "full" returns a named vector of variable importance.}

\item{ntree}{Number of trees to grow. See \link{randomForest}.}

\item{mtry}{Number of predictors randomly sampled as candidates at each
split. See \link{randomForest}.}

\item{...}{Optional arguments passed to \link{randomForest}.}
}
\value{
Ordered vector of names of filtered parameters. If \code{return} is
\code{"full"} a named vector of variable importance is returned.
}
\description{
Fits a random forest model and ranks variables by variable importance.
}
\details{
This filter uses the \link{randomForest} function from the randomForest package.
Variable importance is calculated using the \link{importance} function, specifying
type 1 = mean decrease in accuracy. See \link{importance}.
}
