% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim.R
\name{slim}
\alias{slim}
\title{Slim nestedcv models}
\usage{
slim(x)
}
\arguments{
\item{x}{A 'nestedcv' or 'cva.glmnet' fitted model object.}
}
\value{
For 'nestedcv' objects, a list object of the same class but only
containing \code{outer_result}. For 'cva.glmnet' models, only the cv.glmnet
model with the best alpha value is kept. Models for all other values of
alpha are discarded.
}
\description{
Slims nestedcv objects to only the models in the outer CV folds.
}
\seealso{
\code{\link[=nestcv.glmnet]{nestcv.glmnet()}} \code{\link[=nestcv.train]{nestcv.train()}} \code{\link[=outercv]{outercv()}} \code{\link[=cva.glmnet]{cva.glmnet()}}
}
