% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innercv_roc.R
\name{train_roc}
\alias{train_roc}
\title{Build ROC curve from outer CV training folds}
\usage{
train_roc(x, direction = "<", ...)
}
\arguments{
\item{x}{a \code{nestcv.glmnet}, \code{nestcv.train} or \code{outercv} object}

\item{direction}{Set ROC directionality \link[pROC:roc]{pROC::roc}}

\item{...}{Other arguments passed to \link[pROC:roc]{pROC::roc}}
}
\value{
\code{"roc"} object, see \link[pROC:roc]{pROC::roc}
}
\description{
Build ROC (receiver operating characteristic) curve from outer training
folds. Object can be plotted using \code{plot()} or passed to functions
\code{\link[pROC:auc]{pROC::auc()}} etc.
}
\details{
Note: the argument \code{outer_train_predict} must be set to \code{TRUE} in
the original call to either \code{nestcv.glmnet}, \code{nestcv.train} or \code{outercv}.
}
