% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innercv_roc.R
\name{train_summary}
\alias{train_summary}
\title{Summarise performance on outer training folds}
\usage{
train_summary(x)
}
\arguments{
\item{x}{a \code{nestcv.glmnet} object}
}
\value{
Returns performance metrics from outer training folds, see
\link{predSummary}.
}
\description{
Calculates performance metrics on outer training folds: confusion matrix,
accuracy and balanced accuracy for classification; ROC AUC for binary
classification; RMSE for regression.
}
\details{
Note: currently only works for \code{nestcv.glmnet} objects. The argument
\code{outer_train_predict} must be set to \code{TRUE} in the original nested CV call.
}
\seealso{
\link{predSummary}
}
