% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionality.R
\name{var_direction}
\alias{var_direction}
\title{Variable directionality}
\usage{
var_direction(object)
}
\arguments{
\item{object}{a \code{nestcv.glmnet} or \code{nestcv.train} fitted model}
}
\value{
named vector showing sign of final predictors. If the response vector
is multinomial \code{NULL} is returned.
}
\description{
Determines directionality of final predictors for binary or regression
models.
}
