% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_filter_balance.R
\name{weight}
\alias{weight}
\title{Calculate weights for class imbalance}
\usage{
weight(y)
}
\arguments{
\item{y}{Response vector}
}
\value{
Vector of weights
}
\description{
Calculate weights for class imbalance
}
