% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{wilcoxon_filter}
\alias{wilcoxon_filter}
\title{Wilcoxon test filter}
\usage{
wilcoxon_filter(
  y,
  x,
  nfilter = NULL,
  p_cutoff = 0.05,
  type = c("index", "names", "full"),
  ...
)
}
\arguments{
\item{y}{Response vector}

\item{x}{Matrix of predictors}

\item{nfilter}{Number of predictors to return. If \code{NULL} all predictors with
p values < \code{p_cutoff} are returned.}

\item{p_cutoff}{p value cut-off}

\item{type}{Type of vector returned. Default "index" returns indices,
"names" returns predictor names, "full" returns a matrix of p values.}

\item{...}{args}
}
\value{
Integer vector of indices of filtered parameters (type = "index") or
character vector of names (type = "names") of filtered parameters. If
\code{type} is \code{"full"} full output from \link[matrixTests:wilcoxon]{matrixTests::row_wilcoxon_twosample}
is returned.
}
\description{
Simple univariate filter using Wilcoxon test using the matrixTests package.
}
